requirements.txt
##################################
#           PRODUCTION           #
##################################

##### Frameworks #####
autonomous-app >= 0.1.53

##### Configuration and Documentation #####


##### System #####


##### API #####


##### Database #####


##### Frontend #####


##### Network #####

requests
gunicorn

##### Security #####


##### Async ##### 

##### AI #####


##### Parsing #####

pypandoc
dateparser
markdown-to-json

##### Utility #####

d20

##################################
#           DEVELOPMENT          #
##################################

pytest
pytest-cov



